+++ src/helpers/manage_posted_data.py
@@ -17,7 +17,7 @@ FILE_STRUCTURE = {
     "zero_days": [],
     "details_available_soon": [],
     "posts": {
-        "new_updates": [],
+        "new_releases": [],
         "new_sec_content": [],
         "ios_modules": [],
         "zero_days": {},
@@ -37,12 +37,14 @@ def read() -> dict:
 
     assert posted_data_json["posts"]["new_sec_content"] != [], "ERROR: 'new_sec_content' list inside of posted_data.json is empty. This is used to recognize that releases published before them are new. Add at least last 3 release names there."
 
+    assert posted_data_json["posts"]["new_releases"] != [], "ERROR: 'new_releases' list inside of posted_data.json is empty. This is used to recognize that releases published before them are new. Add at least last 3 release names there."
+
     return posted_data_json
 
 
 def clear_old_data(new_data: dict) -> dict:
-    while len(new_data["posts"]["new_updates"]) > 15:
-        new_data["posts"]["new_updates"].pop(0)
+    while len(new_data["posts"]["new_releases"]) > 15:
+        new_data["posts"]["new_releases"].pop(0)
 
     while len(new_data["posts"]["new_sec_content"]) > 15:
         new_data["posts"]["new_sec_content"].pop(0)